src/main/java/com/example/JEE/controllers/ReviewController.java
package com.example.JEE.controllers;

import com.example.JEE.entities.Review;
import com.example.JEE.services.ReviewService;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
@RequestMapping("/reviews")
public class ReviewController {

    @Autowired
    private ReviewService reviewService;

    @PostMapping
    public Review createReview(@RequestBody Review review) {
        if (review.getRating() <= 10) {
            return reviewService.createReview(review);
        } else {
            return null;  // Return null if the rating is greater than 10
        }
    }

    @GetMapping("/{id}")
    public ResponseEntity<Review> getReviewById(@PathVariable int id) {
        Optional<Review> review = reviewService.getReviewById(id);
        if (review.isPresent()) {
            return ResponseEntity.ok(review.get());  // Return review with 200 OK
        } else {
            return ResponseEntity.status(404).build();  // Return 404 NOT FOUND if not found
        }
    }

    @GetMapping
    public List<Review> getAllReviews() {
        return reviewService.getAllReviews();
    }

    @DeleteMapping("/{id}")
    public void deleteReview(@PathVariable int id) {
        reviewService.deleteReview(id);
    }
}
